Fixed so that "CheckIfPlayerIsSeenForEnemy" and "GetClosestEnemy" works properly. Probably. Maybe?

+++ Radiant/Radiant/OptionsState.cpp
@@ -64,7 +64,7 @@ void OptionsState::Init()
 	});
 
 	_builder->Text()->ChangeFontSize(b1, (uint)(fontSize));
-	_builder->Overlay()->SetExtents(b1, _builder->Text()->GetLength(b1), (uint)fontSize);
+	_builder->Overlay()->SetExtents(b1, _builder->Text()->GetLength(b1), fontSize);
 	_controller->ToggleVisible(b1, false);
 	_controller->ToggleEventChecking(b1, false);
 
@@ -80,7 +80,7 @@ void OptionsState::Init()
 
 	});
 	_builder->Text()->ChangeFontSize(b2, (uint)(fontSize));
-	_builder->Overlay()->SetExtents(b2, _builder->Text()->GetLength(b2), (uint)fontSize);
+	_builder->Overlay()->SetExtents(b2, _builder->Text()->GetLength(b2), fontSize);
 
 	// Discard changes text
 	Entity sh = _builder->CreateLabel(
@@ -91,7 +91,7 @@ void OptionsState::Init()
 		45.0f,
 		"");
 	_builder->Text()->ChangeFontSize(sh, (uint)(fontSize));
-	_builder->Overlay()->SetExtents(sh, _builder->Text()->GetLength(sh), (uint)fontSize);
+	_builder->Overlay()->SetExtents(sh, _builder->Text()->GetLength(sh), fontSize);
 
 	Entity byes = _builder->CreateButton(
 		XMFLOAT3(50.0f*widthPercentOfDefault + _builder->Text()->GetLength(sh)+20.0f*heightPercentOfDefault, height - 80.0f*heightPercentOfDefault, 0.0f),
@@ -105,7 +105,7 @@ void OptionsState::Init()
 		ChangeStateTo(StateChange(new MenuState));
 	});
 	_builder->Text()->ChangeFontSize(byes, (uint)(fontSize));
-	_builder->Overlay()->SetExtents(byes, _builder->Text()->GetLength(byes), (uint)fontSize);
+	_builder->Overlay()->SetExtents(byes, _builder->Text()->GetLength(byes), fontSize);
 
 	Entity bno = _builder->CreateButton(
 		XMFLOAT3(50.0f*widthPercentOfDefault + _builder->Text()->GetLength(sh) + 40.0f*heightPercentOfDefault + _builder->Text()->GetLength(byes), height - 80.0f*heightPercentOfDefault, 0.0f),
@@ -118,7 +118,7 @@ void OptionsState::Init()
 
 	});
 	_builder->Text()->ChangeFontSize(bno, (uint)(fontSize));
-	_builder->Overlay()->SetExtents(bno, _builder->Text()->GetLength(bno), (uint)fontSize);
+	_builder->Overlay()->SetExtents(bno, _builder->Text()->GetLength(bno), fontSize);
 
 	_controller->BindEvent(bno, EventManager::EventType::LeftClick,
 		[b2, sh,bno,byes,a,c]() {
